Replaced OOP to POP
